LayoutTests/crypto/subtle/ecdh-import-key-derive-bits-custom-length-expected.txt

Test ECDH deriveBits operation with imported base key and custom length

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bytesToHexString(derivedKey) is expectedDerivedKey
PASS successfullyParsed is true

TEST COMPLETE
